remove obsolete old-style CAN table generation
